+++ src/main/java/com/invadermonky/fancyoutlines/config/ConfigTags.java
@@ -4,9 +4,12 @@ import com.invadermonky.fancyoutlines.FancyOutlines;
 import com.invadermonky.fancyoutlines.utils.HighlighterHolder;
 import net.minecraft.block.Block;
 import net.minecraft.block.state.IBlockState;
+import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.init.Blocks;
 import net.minecraft.util.ResourceLocation;
+import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.MathHelper;
+import net.minecraftforge.common.ForgeHooks;
 import net.minecraftforge.fml.common.registry.ForgeRegistries;
 
 import javax.annotation.Nullable;
@@ -17,15 +20,24 @@ import java.util.regex.Pattern;
 
 public class ConfigTags {
     private static HighlighterHolder GLOBAL_HIGHLIGHTER;
+    private static HighlighterHolder NO_HARVEST_HIGHLIGHTER;
     private static List<HighlighterHolder> BLOCK_HIGHLIGHTS = new ArrayList<>();
 
     @Nullable
-    public static HighlighterHolder getBlockHighlight(IBlockState state) {
+    public static HighlighterHolder getBlockHighlight(EntityPlayer player, IBlockState state, BlockPos pos) {
+        if(ConfigHandlerBH.noHarvestEnable && !canPlayerHarvest(player, state, pos)) {
+            return NO_HARVEST_HIGHLIGHTER;
+        }
         return BLOCK_HIGHLIGHTS.stream().filter(highlighter -> highlighter.matches(state)).findAny().orElse(GLOBAL_HIGHLIGHTER);
     }
 
+    public static boolean canPlayerHarvest(EntityPlayer player, IBlockState state, BlockPos pos) {
+        return ForgeHooks.canHarvestBlock(state.getBlock(), player, player.world, pos);
+    }
+
     public static void syncConfig() {
-        GLOBAL_HIGHLIGHTER = new HighlighterHolder();
+        GLOBAL_HIGHLIGHTER = HighlighterHolder.getDefaultHolder();
+        NO_HARVEST_HIGHLIGHTER = HighlighterHolder.getNoHarvestHolder();
         BLOCK_HIGHLIGHTS.clear();
 
         Pattern pattern = Pattern.compile("^([^:]+:[^:\\s]+):?(\\d+)?=(\\d+);(#[0-9a-fA-F]+|\\d+)$");